+++ cli/index.js
@@ -7,9 +7,9 @@ const args = require('minimist')(process.argv.slice(2));
 const request = require('request-promise-native');
 const chalk = require('chalk');
 
-const { getBatchProps } = require('./batch');
+const { getBatchProps, getBatchIssue } = require('./batch');
 const parseAndValidateInput = require('./input');
-const { capitalize, compareText, uniq } = require('./utils');
+const { compareText, getProjectName, getGraph } = require('./utils');
 
 const snykBaseUrl = 'https://snyk.io/api/v1';
 
@@ -189,8 +189,7 @@ async function createIssues() {
     const issueQuestions = [];
 
     let ctr = 0;
-    console.log(`Found ${issues.length} vulnerabilities:
-`);
+    console.log(`Found ${issues.length} vulnerabilities:\n`);
     issues.forEach((issue, i) => {
         const description = `${i + 1}. ${issue.package} ${issue.version} - ${
             issue.title
@@ -217,28 +216,6 @@ ${getGraph(project, issue, ' * ')}
     await generateGhIssues(project, issuesToAction);
 
     process.exit(0);
-}
-
-function getProjectName(project) {
-    return typeof args.projectName !== 'undefined'
-        ? args.projectName
-        : project.name;
-}
-
-function getManifestName(project) {
-    if (args.parseManifestName) {
-        return project.name.substring(project.name.indexOf(':') + 1);
-    }
-    return getProjectName(project);
-}
-
-function getGraph(project, issue, prefix) {
-    return issue.from
-        .map(
-            (paths) =>
-                `${prefix}${getManifestName(project)} > ${paths.join(' > ')}`
-        )
-        .join('\r\n');
 }
 
 // Must include Snyk id to distinguish between issues within the same component, that might have different mitigation
@@ -268,66 +245,17 @@ async function generateGhIssues(project, issues, existingMap = new Map()) {
         typeof args.ghLabels !== 'undefined' ? args.ghLabels.split(',') : [];
     labels.push('snyk');
 
-    const projectName = getProjectName(project);
     let ghNewIssues = [];
     let ghUpdatedIssues = [];
     if (batch && issues.length) {
-        const sevMap = issues.reduce((acc, cur) => {
-            acc[cur.severity] = (acc[cur.severity] || []).concat(cur);
-            return acc;
-        }, {});
-        const severities = Object.keys(sevMap)
-            .map((sev) => capitalize(sev))
-            .join(`/`);
-        const batchProps = await getBatchProps(issues);
-
-        // if there is a single vulnerability, just use that for the description
-        let description = `${issues[0].title}`;
-        if (issues.length > 1) {
-            // otherwise, if there are multiple vulnerabilities:
-            const titles = uniq(issues.map((x) => x.title));
-            const vuln = titles.length === 1 ? ` ${titles[0]}` : '';
-            // if there are multiple of vulnerabilities of a single type, include the type in the description
-            // otherwise, if there are multiple types of vulnerabilities of a multiple types, leave that out of the description
-            description = `${issues.length}${vuln} findings`;
-        }
-        const title = `${getProjectName(project)} - ${description} in ${
-            batchProps.package
-        } ${batchProps.version} (${severities})`;
-
-        const text = Object.keys(sevMap)
-            .map((sev) => {
-                const header = `# ${capitalize(sev)}-severity vulnerabilities`;
-                const body = sevMap[sev]
-                    .map(
-                        (issue, i) =>
-                            `\r\n\r\n<details>
-<summary>${i + 1}. ${issue.title} in ${issue.package} ${issue.version} (${
-                                issue.id
-                            })</summary>
-
-## Detailed paths
-${getGraph(project, issue, '* ')}
-
-${issue.description}
-- [${issue.id}](${issue.url})
-</details>`
-                    )
-                    .join('');
-                return header + body;
-            })
-            .join('\r\n\r\n');
+        const { title, body } = await getBatchIssue(project, issues);
 
         ghNewIssues = [
             await octokit.issues.create({
                 owner: ghOwner,
                 repo: ghRepo,
                 title,
-                body: `This issue has been created automatically by a source code scanner.
-
-Snyk project: [\`${project.name}\`](${project.browseUrl}) (manifest version ${project.imageTag})
-
-${text}`,
+                body,
                 labels,
             }),
         ];
